Palm Reader/requirements.txt
streamlit
python-dotenv
requests
pillow